+++ UKHO.PeriodicOutputService/UKHO.BESS.API.FunctionalTests/Models/ResponseBatchStatusModel.cs
@@ -3,7 +3,6 @@ namespace UKHO.ExchangeSetService.API.FunctionalTests.Models
 {
     public class ResponseBatchStatusModel
     {
-        public string? BatchId { get; set; }
         public string? Status { get; set; }
     }
 }
